feat(wallet): Implement wallet factory and service worker integration
- Added wallet factory to create wallet implementations based on execution context (SERVICE_WORKER, WORKER_CLIENT, DIRECT).
- Introduced service worker initialization utilities for automatic wallet setup.
- Created a bundled service worker to handle wallet operations and communication with the main thread.
- Implemented message handling for various wallet operations including initialization, address retrieval, balance checking, and transaction signing.
- Added error handling and response mechanisms for service worker messages.
