+++ src/identity/index.ts
@@ -11,3 +11,4 @@ export interface Identity {
 
 export * from "./singleKey";
 export * from "./serviceWorker";
+export * from "./proxyIdentity";
